+++ src/app/component-registry.js
@@ -11,6 +11,8 @@ import LayerControllerSubscriber from '../components/layer-controller/LayerContr
 import HiGlassSubscriber from '../components/higlass/HiGlassSubscriber';
 import CellSetSizesPlotSubscriber from '../components/sets/CellSetSizesPlotSubscriber';
 import GenomicProfilesSubscriber from '../components/higlass/GenomicProfilesSubscriber';
+import ExpressionHistogramSubscriber from '../components/genes/ExpressionHistogramSubscriber';
+import CellSetExpressionPlotSubscriber from '../components/sets/CellSetExpressionPlotSubscriber';
 
 
 const registry = {
@@ -24,6 +26,8 @@ const registry = {
   [Component.LAYER_CONTROLLER]: LayerControllerSubscriber,
   [Component.CELL_SET_SIZES]: CellSetSizesPlotSubscriber,
   [Component.GENOMIC_PROFILES]: GenomicProfilesSubscriber,
+  [Component.EXPRESSION_HISTOGRAM]: ExpressionHistogramSubscriber,
+  [Component.CELL_SET_EXPRESSION]: CellSetExpressionPlotSubscriber,
   // The plain higlass component does not abstract away the HiGlass view config,
   // so we probably want to avoid documenting it, only use it for development purposes.
   higlass: HiGlassSubscriber,