+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ### Added
 - Added a new documentation site.
+- Added an optional `description` field to the dataset definition config object. Updated the `<Description/>` component to prefer this value over the top-level description value, if available.
 
 ### Changed
+- Fix AnnData text decoding.
+- Refactor AnnData flat array decoding and resolve bug.
 
 ## [1.1.3](https://www.npmjs.com/package/vitessce/v/1.1.3) - 2021-01-07
 
@@ -34,6 +37,7 @@
 - Support loading AnnData from zarr.
 - Added scatterplot component options pane.
 - Added a `TextLayer` to render cell set names over the corresponding cell set centroid in the scatterplot.
+- Added a gene expression histogram Vega-Lite component and a cell set by gene expression violin plot Vega component.
 
 ### Changed
 - Genes list is now a virtual scroll to allow for more performant rendering.